Fix format_euros()
This fixes two related issues:
- Total amount reaching 327,67 EUR does no longer result in negativ numbers
- Total amount less than 0 EUR gets calculated and displayed correctly (but
  not across reloads yet)

fixes #36

+++ include/config.h
@@ -9,8 +9,8 @@
 #define CREDITS_PER_PAGE 10
 
 /* Eingenommes Geld in Cent */
-extern unsigned long int money;
-extern unsigned long int items_sold;
+extern int32_t money;
+extern int32_t items_sold;
 extern BYTE printer_port;
 
 /* Datenstruktur der verkauften Einträge */
